requirements.txt
djangorestframework>=3.12.4,<3.13
django>=3.2.2,<3.3
flake8>=3.9.1,<3.10
psycopg2-binary>=2.8.6,<2.9
